move bookmark filter out of search filter. move preprocessor for search filter into its own file
